% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_map_core.R
\name{IO_map}
\alias{IO_map}
\title{Plots a (ggplot) map of the IO and (optionally) an overlay of data visualization artifacts}
\usage{
IO_map(
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  fill_color = "black",
  border_color = "black",
  show_grids = TRUE,
  show_IO_areas = TRUE,
  show_EEZs = FALSE,
  show_high_seas = TRUE,
  content_drawer = NULL,
  draw_content_first = TRUE
)
}
\arguments{
\item{xlim}{A vector with the min - max longitude values}

\item{ylim}{A vector with the min - max latitude values}

\item{fill_color}{The fill color used to draw the map}

\item{border_color}{The border color used to draw the map}

\item{show_grids}{Flag to display reference grids every 20 degrees (both horizontally and vertically)}

\item{show_IO_areas}{Flag to display the two main IO areas (F51 and F57) as a separate layer}

\item{show_EEZs}{Flag to display the boundary lines of the EEZs}

\item{show_high_seas}{Flag to display the boundary lines of the high seas}

\item{content_drawer}{an (optional) function to plot an overlay of data visualization artifacts}

\item{draw_content_first}{Flag to force plotting the content (through \code{content_drawer} when provided) before or after the other layers (grids, IO areas and EEZs)}
}
\value{
A map with the features enabled by the flags and parameters above, with an optional overlay of data visualization artifacts of any kind
}
\description{
Plots a (ggplot) map of the IO and (optionally) an overlay of data visualization artifacts
}
