% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_map_constants.R
\docType{data}
\name{IO_map_ylim}
\alias{IO_map_ylim}
\title{Default latitude limits for the IO map}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
IO_map_ylim
}
\description{
Default latitude limits for the IO map
}
\keyword{datasets}
